LayoutTests/imported/w3c/web-platform-tests/html/rendering/non-replaced-elements/the-fieldset-and-legend-elements/min-inline-size-expected.txt





PASS horizontal-tb
FAIL vertical-lr  assert_equals: width expected "0px" but got "100px"
FAIL vertical-rl  assert_equals: width expected "0px" but got "100px"
PASS horizontal-tb override
FAIL vertical-lr override assert_equals: width expected "0px" but got "100px"
FAIL vertical-rl override assert_equals: width expected "0px" but got "100px"
